extension: Wire up functional checker
